scripts/requirements.txt
redis==5.0.1
pymongo==4.6.2
neo4j==5.22.0
elasticsearch==8.13.0
psycopg2-binary==2.9.9
faker==26.3.0
python-dotenv==1.0.1 